Add class `KerasDirectory`
This class takes a `ChipDataset` and hard links the files to a new
directory. The new directory is set up such that each class for training
is in its own directory, as Keras expects. This will make it possible to
easily "stream" data from disk for training (instead of loading it all
into memory).