Change how `set_type` is set in `ChipDataset()`
Previously it was set with an integer from the `SetType` enum. It is now
set with the enum itself, or a string describing the set ("train",
"test", etc.). `None` is now the default value in the constructor, which
defaults to `SetType.ALL`. All checks have been changed to assume that
an acutal `SetType` object is stored in the class.