Add color and vehicle type to VeRI chips (#66)
* Change VeRI ChipDataset to use `os.path.join`

* Add `color` and `vehicle_type` to VeRI data

This still needs tests.

* Switch the VeRI test to use "TRAIN"

Previously it was reading a fake "test" set. However, the training set
will soon have a superset of the functionality of the test set.

* Fix some bugs exposed in VeRI

* Add tests for VeRI's `color` and `vehicle_type`

* Change `test_veri.py` to use `os.path.join()`

* Allow VeRI to throw if it fails

The VeRI dataset no longer silently fails, it will now throw errors when
construction it's `self.__color_type` dictionary.

Additionally, the 0 ID for colors and body types has been added. Some
body types in the training XML are set to 0 when they are not one of
predefined types.

* Change VeRI vehicle ID to int

We assume the vehicle ID is an int elsewhere, so set it to an int here.

* Prefer `{}` to `dict()`

* Add more helpful error message for #72

* Clean up some PEP8 issues

* Add patch to help workaround #76 in `VeriDataset`